Made RAHT execute directly with fixed point

+++ haar3D.h
@@ -15,7 +15,7 @@ int64_t sqrtIF(int64_t A, int64_t W0, int64_t W1);
 int64_t sqrtIF(int64_t A, int64_t W);
 
 void copyAsort(double *VX, double *CX, size_t N, double *C, uint64_t *W, uint64_t *val, uint64_t *ord);
-void haar3D(double Qstep, double *inV, double *inC, size_t K, size_t N, size_t depth, double *outCT, double *outW=NULL);
-void inv_haar3D(double Qstep, double *inV, double *inCT, size_t K, size_t N, size_t depth, double *outC);
+void haar3D(int64_t Qstep, double *inV, double *inC, size_t K, size_t N, size_t depth, double *outCT, double *outW=NULL);
+void inv_haar3D(int64_t Qstep, double *inV, double *inCT, size_t K, size_t N, size_t depth, double *outC);
 
 #endif // EXTRA_H
